feat: local compiler mode for dev, release version pinning
