+++ .github/workflows/build.yaml
@@ -32,8 +32,21 @@ jobs:
     - name: Build
       run: dotnet build -c Release --no-restore
       
-    - name: Run tests
+    - name: Run .NET tests
       run: dotnet test -c Release --no-build --verbosity normal
+
+    - name: Setup Node.js
+      uses: actions/setup-node@v4
+      with:
+        node-version: '20.x'
+
+    - name: Install extension dependencies
+      working-directory: extension
+      run: npm ci
+
+    - name: Run extension tests
+      working-directory: extension
+      run: npm test
 
   build:
     needs: test
